+++ src/main/java/com/gerenciamentodeusuarioapirest/controller/PerfilController.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,6 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.gerenciamentodeusuarioapirest.models.Cargo;
 import com.gerenciamentodeusuarioapirest.models.Perfil;
 import com.gerenciamentodeusuarioapirest.repositorys.PerfilRepository;
+import com.gerenciamentodeusuarioapirest.utilidade.SimplesResposta;
 
 @CrossOrigin(origins = "*")
 @RestController
@@ -35,11 +37,17 @@ public class PerfilController {
 		return perfilRepository.save(perfil);
 	}
 
-	@DeleteMapping("/perfis")
-	public void deletaPerfil(@RequestBody @Valid Perfil perfil) {
-		perfilRepository.delete(perfil);
+	@DeleteMapping("/perfis/{id}")
+	public SimplesResposta deletaPerfil(@PathVariable(value = "id") long id) {
+		Perfil perfil = perfilRepository.findById(id);
+		if (!perfil.verificarSeHaUsuarioVinculado()) {
+			perfilRepository.delete(perfil);
+			return new SimplesResposta("Perfil deletado com sucesso!");
+		} else {
+			return new SimplesResposta("Perfil não pode ser deletado por possui usuário vinculado");
+
+		}
+
 	}
-
-
 
 }
